Change version number to 0.1.0
During this phase of development, I will be tweaking the output of the tool a lot. If I started at 1.0.0 and followed SemVer, every rearrangement of existing fields in the JSON, or any new command-line flags, would require a major version bump. Remembering to bump the version every time would be annoying, and a version number of v20 might give the wrong idea about the stability of this tool. So I’ll just keep the version 0.x unless and until I make the tool parse every field I think it will ever parse and stabilize the command-line API.