Create a `tsconfig.json`
I created it by running `tsc --init`. I started to reformat it with simplified comment syntax in the hopes that Prettier would format it better, but it didn’t help, so I told Prettier to ignore that file.

This file will be necessary to compile the TypeScript to JavaScript for publishing on npm.