Disable esModuleInterop in TypeScript config to fix Yargs-related bug
Before this fix, `yarn start --help`, which uses Yargs’s auto-constructed help output, would only print the following:

~~~
Options:
  --help     Show help                                                 [boolean]
  --version  Show version number                                       [boolean]
~~~

After this fix the output is correct again:

~~~
Usage: index.ts [option] [file]

Options:
  --format, -f  Data format to output
      [choices: "generic", "audacity_label_track", "debug"] [default: "generic"]
  --version     Show version number                                    [boolean]
  --help        Show help                                              [boolean]

Examples:
  index.ts myTranscribeFile.xsc             reading input with a file argument
  cat myTranscribeFile.xsc | index.ts >     converting input with stdin and
  transcribeFileData.json                   stdout redirection
~~~

I found that this flag was the cause only through trial and error, after failing to find the cause in my code or in the versions of the packages I was using.

Reading about the esModuleInterop flag on https://stackoverflow.com/q/56238356/578288, I can see that this flag was the problem because the Yargs library defines a builder that mutates the root yargs object, the object it exports. The esModuleInterop flag tells TypeScript to wrap exported objects like {default: yargs} in some cases, so it must have messed with the state of the exported yargs object that was being mutated.