Support exporting to Audacity label track format
The documentation hasn’t been updated to mention this yet. The program does still hint that this is possible when run with the `--help` flag.

I manually compiled the program and used it to create a label track, and Audacity was able to import it correctly.

The ignored folder “tests/Audacity label import tests” contains label track text files and Audacity projects that I used to manually check the behavior of label track importing in Audacity.