Add a manually-generated table of contents
I’m adding this just for the “Installation and usage” link, so the user will be able to get help quicker if they see “transcribe-xsc-convert” at the top, try to install it with npm, and get an error.

This is kind of heavyweight and I might replace it with just a single link to “Installation and usage”. But since I already made it, I’d like to commit it so it’s easy to get back later.